+++ Bloch/docs/release-notes.md
@@ -5,7 +5,7 @@
 ## Upcoming release
 
 * Added the Bloch.EXE card.
-* Fixed Veiling Aura armored/weak part calculation.
+* Fixed Veiling Aura armored/weak/brittle part calculation.
 
 ## 1.0.0
 Released 2 June 2024.